requirements.txt
grpcio == 1.3.5
grpcio-tools == 1.3.5
protobuf == 3.5.1
leveldb == 0.20
coloredlogs == 9.0
verboselogs == 1.7
cryptography == 2.2.2
PyJWT == 1.5.3
pyopenssl == 17.5.0
yappi == 0.98
GitPython == 2.1.8
requests == 2.19.1
fluent-logger == 0.9.1
gunicorn == 19.7.1
setproctitle == 1.1.10
python-pkcs11 == 0.5.0
terminaltables == 3.1.0
pysha3 == 1.0.2
secp256k1 == 0.13.2
jsonrpcserver == 3.5.3
jsonrpcclient == 2.6.0
ordered-set == 2.0.2
aiohttp == 3.0.9
sanic == 0.7.0
aio_pika == 3.0.1
pika == 0.12.0
earlgrey == 0.0.4
